--- src/main/java/payroll/EmployeeRepository.java
@@ -1,10 +1,0 @@
-package payroll;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * 従業員リポジトリ
- */
-interface EmployeeRepository extends JpaRepository<Employee, Long> {
-
-}
