--- src/main/java/payroll/OrderRepository.java
@@ -1,9 +1,0 @@
-package payroll;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * 注文リポジトリ
- */
-interface OrderRepository extends JpaRepository<Order, Long> {
-}
