+++ src/main/java/payroll/assembler/EmployeeModelAssembler.java
@@ -1,15 +1,18 @@
-package payroll;
+package payroll.assembler;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
 
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.server.RepresentationModelAssembler;
 import org.springframework.stereotype.Component;
+import payroll.controller.EmployeeController;
+import payroll.entity.Employee;
 
 /**
  * EmployeeエンティティをEmployeeエンティティモデルに変換するクラス
  */
 @Component
+public
 class EmployeeModelAssembler implements RepresentationModelAssembler<Employee, EntityModel<Employee>> {
     /**
      * Employeeエンティティをモデルに変換するメソッド
