+++ src/main/java/payroll/exception/EmployeeNotFoundAdvice.java
@@ -1,25 +1,26 @@
-package payroll;
+package payroll.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import payroll.exception.EmployeeNotFoundException;
 
 /**
- * 注文検索例外発生時の処理用クラス
+ * 従業員検索エラー時の処理用クラス
  */
 @ControllerAdvice
-public class OrderNotFoundAdvice {
+public class EmployeeNotFoundAdvice {
     /**
-     * 例外メッセージ返却
+     * エラーメッセージの取得
      * @param ex
-     * @return エラーメッセージ
+     * @return
      */
     @ResponseBody
-    @ExceptionHandler(OrderNotFoundException.class)
+    @ExceptionHandler(EmployeeNotFoundException.class)
     @ResponseStatus(HttpStatus.NOT_FOUND)
-    String employeeNotFoundHandler(OrderNotFoundException ex) {
+    String employeeNotFoundHandler(EmployeeNotFoundException ex) {
         return ex.getMessage();
     }
 }
