+++ src/main/java/payroll/exception/EmployeeNotFoundException.java
@@ -1,15 +1,15 @@
-package payroll;
+package payroll.exception;
 
 /**
  * 従業員が見つからない時の処理用クラス
  */
-class EmployeeNotFoundException extends RuntimeException {
+public class EmployeeNotFoundException extends RuntimeException {
 
     /**
      * エラーメッセージ表示
      * @param id
      */
-    EmployeeNotFoundException(Long id) {
+    public EmployeeNotFoundException(Long id) {
         super("Could not find employee " + id);
     }
 }
