+++ src/main/java/payroll/exception/OrderNotFoundException.java
@@ -1,4 +1,4 @@
-package payroll;
+package payroll.exception;
 
 /**
  * 注文未存在例外クラス
@@ -8,7 +8,7 @@ public class OrderNotFoundException extends RuntimeException {
      * コンストラクタ
      * @param id
      */
-    OrderNotFoundException(Long id) {
+    public OrderNotFoundException(Long id) {
         super("Could not find order " + id);
     }
 }
