+++ src/main/java/payroll/repository/EmployeeRepository.java
@@ -1,0 +1,11 @@
+package payroll.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import payroll.entity.Employee;
+
+/**
+ * 従業員リポジトリ
+ */
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+
+}
